split public api for file_watcher from private implementation
