+++ lib/file_watcher.js
@@ -1,292 +1,125 @@
-const path = require('path');
 const EventEmitter = require('events').EventEmitter;
-const { hasMagic } = require('glob');
-const { Minimatch } = require('minimatch');
 
-const isEmfileError = require('./utils/is_emfile_error');
-const { convertToPosix } = require('./utils/posix');
-const {
-  buildWatchGlobPolicy,
-  pathMatchesWatchTarget,
-} = require('./utils/file_watch_glob_policy');
+const Impl = require('./file_watcher_impl');
 
-/** @private Only {@link FileWatcher.create} may construct instances. */
-const kInternal = Symbol('FileWatcher.internal');
-
-function includePatternHasGlobMagic(pattern) {
-  return new Minimatch(pattern, { dot: true }).hasMagic();
-}
-
-function resolvedPathIsUnderCwd(resolved) {
-  const rel = path.relative(process.cwd(), resolved);
-  return rel !== '' && !rel.startsWith('..') && !path.isAbsolute(rel);
-}
+const kCompose = Symbol('FileWatcher.compose');
 
 /**
- * Whether chokidar's `all` event (add/addDir) refers to the same path as `file` / `absTarget`.
+ * Coordinates file watching for Testem: it observes the current working directory (and any
+ * non-glob paths from config that live outside it), applies the same include / ignore rules as
+ * the rest of the app (`src_files`, `src_files_ignore`, `watch_files`, etc.), and reports when
+ * paths that matter for the test run change.
+ *
+ * Listen for {@link FileWatcher#event:fileChanged} to react to changes, and optionally for
+ * {@link FileWatcher#event:EMFILE} when the process hits open-file limits.
+ *
+ * ### Construction
+ *
+ * **Do not call `new FileWatcher(...)`.** The constructor is not part of the public API; it
+ * exists only for internal wiring and will always throw if you invoke it from outside this
+ * module. Obtain instances exclusively via {@link FileWatcher.create} with a config object that
+ * exposes `get(key)` and `isCwdMode()` (the same shape as Testem’s `Config`).
+ *
+ * ### Async vs synchronous
+ *
+ * **Return a `Promise` — use `await`, `.then()`, or `.catch()`:**
+ * {@link FileWatcher.create}, {@link FileWatcher#add}, {@link FileWatcher#close}.
+ *
+ * **Synchronous:** {@link FileWatcher#onFileChanged}, {@link FileWatcher#onEMFILE}, and
+ * inherited `EventEmitter` methods (`on`, `once`, `emit`, `removeListener`, …).
+ *
+ * @hideconstructor
+ * @fires FileWatcher#event:fileChanged
+ * @fires FileWatcher#event:EMFILE
  */
-function dynamicAddEventMatchesPath(watchedPath, file, absTarget) {
-  const absWatched = path.resolve(watchedPath);
-  const relWatched = convertToPosix(
-    path.relative(process.cwd(), absWatched),
-  );
-  const relTarget = convertToPosix(
-    path.relative(process.cwd(), absTarget),
-  );
-  return (
-    absWatched === absTarget ||
-    relWatched === relTarget ||
-    watchedPath === file
-  );
-}
-
 class FileWatcher extends EventEmitter {
   /**
-   * @param {{ get: function(string): *, isCwdMode: function(): boolean }} config
-   * @param {symbol} internalToken
+   * Not for public use. Application code must use {@link FileWatcher.create} instead; calling
+   * `new FileWatcher` directly always throws.
+   *
+   * @private
+   * @param {*} impl Internal implementation instance.
+   * @param {symbol} token Internal composition token (not exposed).
    */
-  constructor(config, internalToken) {
+  constructor(impl, token) {
     super();
-
-    if (internalToken !== kInternal) {
+    if (token !== kCompose) {
       throw new TypeError(
         'FileWatcher is not constructable; use FileWatcher.create(config)',
       );
     }
-
-    const policy = buildWatchGlobPolicy(config);
-    this._watchPolicy = policy;
-    this._explicitWatchRelPaths = new Set();
-    const opts = { ignoreInitial: true };
-    if (policy.ignorePatterns.length) {
-      opts.ignored = policy.ignorePatterns;
-    }
-
-    this.fileWatcher = FileWatcher.createWatcher('.', opts);
-
-    let readySettled = false;
-    this._readyPromise = new Promise((resolve, reject) => {
-      this.fileWatcher.once('ready', () => {
-        readySettled = true;
-        resolve();
-      });
-      this.fileWatcher.once('error', (err) => {
-        if (readySettled || isEmfileError(err)) {
-          return;
-        }
-        readySettled = true;
-        reject(err);
-      });
-    });
-
-    for (let i = 0; i < policy.includePatterns.length; i++) {
-      const pattern = policy.includePatterns[i];
-      if (includePatternHasGlobMagic(pattern)) {
-        continue;
-      }
-      const resolved = path.resolve(process.cwd(), pattern);
-      if (!resolvedPathIsUnderCwd(resolved)) {
-        this.fileWatcher.add(path.normalize(pattern));
-      }
-    }
-
-    const onFileChanged = this.onChokidarFsEvent.bind(this);
-    this.fileWatcher.on('change', onFileChanged);
-    this.fileWatcher.on('add', onFileChanged);
-    this.fileWatcher.on('unlink', onFileChanged);
-    this.fileWatcher.on('unlinkDir', onFileChanged);
-    this.fileWatcher.on('error', (err) => {
-      if (isEmfileError(err)) {
-        this.onEMFILE();
-      }
-    });
+    this._impl = impl;
+    impl.on('fileChanged', (filePath) => this.emit('fileChanged', filePath));
+    impl.on('EMFILE', () => this.emit('EMFILE'));
   }
 
   /**
-   * Creates a file watcher and waits until the underlying chokidar instance is ready.
+   * Create a watcher, wait until the underlying engine has finished its initial setup, and
+   * return a ready-to-use `FileWatcher`. This is the **only** supported way to obtain an
+   * instance.
    *
-   * @param {{ get: function(string): *, isCwdMode: function(): boolean }} config
-   * @returns {Promise<FileWatcher>}
+   * @param {{ get: function(string): *, isCwdMode: function(): boolean }} config Testem-style
+   *   config (typically a `Config` instance): `get` reads keys such as `src_files`,
+   *   `src_files_ignore`, `watch_files`, `file`, `disable_watching`; `isCwdMode` reflects cwd
+   *   mode.
+   * @returns {Promise<FileWatcher>} Resolves with a connected watcher once startup is complete.
+   *   **Await** this promise; if the engine fails during startup, the promise rejects.
    */
   static async create(config) {
-    const fw = new FileWatcher(config, kInternal);
-    await fw._readyPromise;
-    return fw;
-  }
-
-  onChokidarFsEvent(filePath) {
-    const resolved = path.resolve(filePath);
-    const posixRel = convertToPosix(
-      path.relative(process.cwd(), resolved),
-    );
-    if (
-      pathMatchesWatchTarget(posixRel, this._watchPolicy) ||
-      this._explicitWatchRelPaths.has(posixRel)
-    ) {
-      this.onFileChanged(filePath);
-    }
-  }
-
-  onFileChanged(filePath) {
-    this.emit('fileChanged', filePath);
-  }
-
-  onEMFILE() {
-    this.emit('EMFILE');
+    const impl = await Impl.create(config);
+    return new FileWatcher(impl, kCompose);
   }
 
   /**
-   * Used when `add` fails (validation or chokidar): tear down the watcher, then rethrow.
-   * @param {Error} err
-   * @returns {Promise<never>}
-   */
-  async _closeAfterFailedAdd(err) {
-    try {
-      await this.close();
-    } catch {
-      // ignore close errors; primary error is `err`
-    }
-    throw err;
-  }
-
-  /**
-   * `add` is for concrete paths (e.g. a file the server is serving). Glob patterns belong in
-   * config / {@link buildWatchGlobPolicy}, not here.
+   * Request that a **concrete** filesystem path be watched in addition to the configured
+   * patterns (for example, a file the dev server started serving after startup). Glob or
+   * minimatch strings are not accepted here — only real paths; pattern-based includes stay in
+   * config.
    *
-   * @param {string} file
-   * @returns {Promise<void>}
+   * @param {string} file Path relative to the current working directory or absolute, as
+   *   accepted by the engine.
+   * @returns {Promise<void>} Settles when the engine has finished the add request. **Await**
+   *   this promise (or attach `.catch()`) so validation errors, engine errors, and internal
+   *   teardown on failure are observed.
    */
-  async _validateConcreteAddPath(file) {
-    if (typeof file !== 'string') {
-      await this._closeAfterFailedAdd(
-        new TypeError('FileWatcher.add expects a string path'),
-      );
-    }
-    if (file.length === 0) {
-      await this._closeAfterFailedAdd(
-        new TypeError('FileWatcher.add expects a non-empty path'),
-      );
-    }
-    if (hasMagic(file, { magicalBraces: true })) {
-      await this._closeAfterFailedAdd(
-        new TypeError(
-          'FileWatcher.add does not accept glob patterns; pass a concrete file path: ' +
-            file,
-        ),
-      );
-    }
+  add(file) {
+    return this._impl.add(file);
   }
 
   /**
-   * Chokidar's `add()` schedules async work and may surface failures via `error` instead of
-   * throwing. Wait for that work to settle: first matching `add`/`addDir` on `all`, or `error`,
-   * or a timeout when chokidar finishes without an event (e.g. ENOENT — no `error` emitted).
+   * Tear down the watcher: close the engine, release OS watch handles, and remove listeners on
+   * this `FileWatcher`. Safe to call more than once; subsequent calls resolve promptly.
    *
-   * @param {string} file
-   * @returns {Promise<void>}
+   * @returns {Promise<void>} Resolves when shutdown is complete. **Await** if later code must
+   *   run only after handles are released.
    */
-  async _awaitChokidarDynamicAdd(file) {
-    const engine = this.fileWatcher;
-    const absTarget = path.resolve(process.cwd(), file);
-    let onError;
-    let onAll;
-    let timeoutId;
-
-    const cleanup = () => {
-      if (onError) {
-        engine.removeListener('error', onError);
-      }
-      if (onAll) {
-        engine.removeListener('all', onAll);
-      }
-      if (timeoutId !== undefined) {
-        clearTimeout(timeoutId);
-      }
-    };
-
-    await new Promise((resolve, reject) => {
-      let settled = false;
-
-      const finish = (err) => {
-        if (settled) {
-          return;
-        }
-        settled = true;
-        cleanup();
-        if (err) {
-          reject(err);
-        } else {
-          resolve();
-        }
-      };
-
-      onError = (err) => {
-        finish(err);
-      };
-      engine.prependOnceListener('error', onError);
-
-      onAll = (event, watchedPath) => {
-        if (event === 'add' || event === 'addDir') {
-          if (dynamicAddEventMatchesPath(watchedPath, file, absTarget)) {
-            finish();
-          }
-        }
-      };
-      engine.on('all', onAll);
-
-      // When chokidar finishes without an `all` add/addDir for this path (e.g. already under
-      // watch, or ENOENT with no error event), settle after a bounded wait.
-      timeoutId = setTimeout(() => finish(), 1000);
-
-      try {
-        engine.add(file);
-      } catch (err) {
-        finish(err);
-      }
+  close() {
+    return Promise.resolve(this._impl.close()).then(() => {
+      this.removeAllListeners();
     });
   }
 
   /**
-   * @param {string} file
-   * @returns {Promise<void>}
+   * Manually signal that a path changed, for the same listeners as a real `fileChanged` event.
+   * Used in tests and narrow control flows; normal file edits are reported automatically by the
+   * engine.
+   *
+   * @param {string} filePath Path that should be treated as changed.
+   * @returns {void}
    */
-  async add(file) {
-    await this._validateConcreteAddPath(file);
-    const rel = convertToPosix(
-      path.relative(
-        process.cwd(),
-        path.resolve(process.cwd(), file),
-      ),
-    );
-    this._explicitWatchRelPaths.add(rel);
-    try {
-      await this._awaitChokidarDynamicAdd(file);
-    } catch (err) {
-      await this._closeAfterFailedAdd(err);
-    }
+  onFileChanged(filePath) {
+    return this._impl.onFileChanged(filePath);
   }
 
   /**
-   * Stops the underlying watch engine (closes fs.watch handles, etc.).
-   * Safe to call more than once.
+   * Manually emit the `EMFILE` condition (too many open files) for listeners that show user
+   * guidance. The real engine also surfaces this when appropriate.
    *
-   * @returns {Promise<void>}
+   * @returns {void}
    */
-  close() {
-    if (this.fileWatcher) {
-      const engine = this.fileWatcher;
-      this.fileWatcher = null;
-      return Promise.resolve(engine.close()).then(() => {
-        this.removeAllListeners();
-      });
-    }
-    this.removeAllListeners();
-    return Promise.resolve();
+  onEMFILE() {
+    return this._impl.onEMFILE();
   }
 }
-
-FileWatcher.createWatcher = function createWatcher(rootPath, opts) {
-  return require('chokidar').watch(rootPath, opts);
-};
 
 module.exports = FileWatcher;
