+++ tests/file_watcher_integration_tests.js
@@ -13,12 +13,14 @@ const expect = require('chai').expect;
 const sinon = require('sinon');
 
 const FileWatcher = require('../lib/file_watcher');
+const { getWatchEngine } = require('./support/file_watcher_test_access');
 
 /** Node does not expose a public watcher count; used only to assert no fs.watch leaks. */
 function countActiveFsWatchers() {
-  return process._getActiveHandles().filter(
-    (h) => h && h.constructor && h.constructor.name === 'FSWatcher',
-  ).length;
+  return process
+    ._getActiveHandles()
+    .filter((h) => h && h.constructor && h.constructor.name === 'FSWatcher')
+    .length;
 }
 
 function makeConfig(overrides) {
@@ -63,20 +65,20 @@ function onceEvent(emitter, event, timeoutMs) {
   });
 }
 
-describe('FileWatcher integration', function() {
+describe('FileWatcher integration', function () {
   this.timeout(15000);
 
   let prevCwd;
   let tmpDir;
   let fw;
 
-  beforeEach(function() {
+  beforeEach(function () {
     prevCwd = process.cwd();
     tmpDir = fs.mkdtempSync(path.join(os.tmpdir(), 'testem-fw-int-'));
     fw = null;
   });
 
-  afterEach(async function() {
+  afterEach(async function () {
     if (fw) {
       try {
         await fw.close();
@@ -100,7 +102,7 @@ describe('FileWatcher integration', function() {
     }
   });
 
-  it('emits fileChanged when a new file matching src_files is created', async function() {
+  it('emits fileChanged when a new file matching src_files is created', async function () {
     process.chdir(tmpDir);
     fw = await FileWatcher.create(makeConfig({ src_files: ['*.js'] }));
 
@@ -113,7 +115,7 @@ describe('FileWatcher integration', function() {
     expect(path.basename(filepath)).to.equal('new.js');
   });
 
-  it('emits fileChanged when an existing matching file is modified', async function() {
+  it('emits fileChanged when an existing matching file is modified', async function () {
     fs.writeFileSync(path.join(tmpDir, 'existing.js'), 'v1');
     process.chdir(tmpDir);
 
@@ -128,7 +130,7 @@ describe('FileWatcher integration', function() {
     expect(path.basename(filepath)).to.equal('existing.js');
   });
 
-  it('does not emit fileChanged for paths matching src_files_ignore', async function() {
+  it('does not emit fileChanged for paths matching src_files_ignore', async function () {
     fs.mkdirSync(path.join(tmpDir, 'vendor'), { recursive: true });
     process.chdir(tmpDir);
 
@@ -153,7 +155,7 @@ describe('FileWatcher integration', function() {
     expect(received).to.equal(null);
   });
 
-  it('add forwards a path so later changes to that file are observed', async function() {
+  it('add forwards a path so later changes to that file are observed', async function () {
     fs.mkdirSync(path.join(tmpDir, 'lib'), { recursive: true });
     fs.writeFileSync(path.join(tmpDir, 'lib', 'extra.js'), '');
     process.chdir(tmpDir);
@@ -174,20 +176,20 @@ describe('FileWatcher integration', function() {
     expect(path.basename(filepath)).to.equal('extra.js');
   });
 
-  it('close clears the wrapper reference to the engine (safe to call after use)', async function() {
+  it('close clears the wrapper reference to the engine (safe to call after use)', async function () {
     process.chdir(tmpDir);
     fw = await FileWatcher.create(makeConfig({ src_files: ['*.js'] }));
 
     await delay(400);
 
-    expect(fw.fileWatcher).to.be.ok();
+    expect(getWatchEngine(fw)).to.be.ok();
 
     await fw.close();
 
-    expect(fw.fileWatcher).to.equal(null);
+    expect(getWatchEngine(fw)).to.equal(null);
   });
 
-  it('releases fs.watch handles after add() throws EMFILE (no watcher leak)', async function() {
+  it('releases fs.watch handles after add() throws EMFILE (no watcher leak)', async function () {
     process.chdir(tmpDir);
     fs.writeFileSync(path.join(tmpDir, 'first.js'), 'export {}');
 
@@ -200,7 +202,7 @@ describe('FileWatcher integration', function() {
 
     const sandbox = sinon.createSandbox();
     const originalAdd = fw.add.bind(fw);
-    sandbox.stub(fw, 'add').callsFake(async function(file) {
+    sandbox.stub(fw, 'add').callsFake(async function (file) {
       if (file === 'second.js') {
         const err = new Error('EMFILE: too many open files, watch');
         err.code = 'EMFILE';
